Changed result type to long for fixing calculation error

+++ src/Main.java
@@ -6,7 +6,7 @@ public class Main {
         System.out.printf("Hello and welcome!\n\n");
         System.out.print("Input a number of iterations: ");
         int j = in.nextInt();
-        int x  = 1;
+        long x  = 1;
         for (int i = 1; i <= j; i++) {
             x = x*i;
             System.out.println(i + "! = " + x);
